Fixed game, added card stacking, visual changes
Game now resets properly. It also has card stacking, to ensure that the
window does not need to resize to fit all cards. To do this switched
from initially using PhotoImage objects to Image objects. Changed
to using .pack for the highest level of elements so they stay centered.
Added a win tracker and display on reset game to show dealer and player
win counts.